move MemberKeys class into forms.py
