+++ geoh5py/ui_json/__init__.py
@@ -19,8 +19,5 @@
 # flake8: noqa
 
 from .input_file import InputFile
-from .members import MemberKeys
 from .utils import monitored_directory_copy
 from .validation import InputValidation
-
-MEMBER_KEYS = MemberKeys()
