+++ geoh5py/ui_json/forms.py
@@ -20,7 +20,6 @@ from __future__ import annotations
 from typing import Any
 
 from geoh5py.shared.exceptions import AggregateValidationError, BaseValidationError
-from geoh5py.ui_json import MEMBER_KEYS
 from geoh5py.ui_json.enforcers import EnforcerPool, ValueEnforcer
 from geoh5py.ui_json.parameters import (
     BoolParameter,
@@ -32,6 +31,50 @@ from geoh5py.ui_json.parameters import (
     StringParameter,
     UUIDParameter,
 )
+
+
+class MemberKeys:
+    """Converts in and out of camel (ui.json) and snake (python) case"""
+
+    camel_to_snake: dict[str, str] = {
+        "groupOptional": "group_optional",
+        "dependencyType": "dependency_type",
+        "groupDependency": "group_dependency",
+        "groupDependencyType": "group_dependency_type",
+        "lineEdit": "line_edit",
+        "choiceList": "choice_list",
+        "fileDescription": "file_description",
+        "fileType": "file_type",
+        "fileMulti": "file_multi",
+        "meshType": "mesh_type",
+        "dataType": "data_type",
+        "dataGroupType": "data_group_type",
+        "isValue": "is_value",
+    }
+
+    @property
+    def snake_to_camel(self) -> dict[str, str]:
+        """Gives the inverse map to camel_to_snake."""
+        return {v: k for k, v in self.camel_to_snake.items()}
+
+    def _map_single(self, key: str, convention: str = "snake"):
+        """Map a string from snake to camel or vice versa."""
+
+        if convention == "snake":
+            out = self.camel_to_snake.get(key, key)
+        elif convention == "camel":
+            out = self.snake_to_camel.get(key, key)
+        else:
+            raise ValueError("Convention must be 'snake' or 'camel'.")
+
+        return out
+
+    def map(self, collection: dict[str, Any], convention="snake"):
+        """Map a dictionary from snake to camel or vice versa."""
+        return {self._map_single(k, convention): v for k, v in collection.items()}
+
+
+MEMBER_KEYS = MemberKeys()
 
 
 class ValueAccess:
